+++ qiskit_braket_provider/providers/__init__.py
@@ -22,4 +22,4 @@ Provider classes and functions
 
 from .braket_backend import AWSBraketBackend, BraketLocalBackend
 from .braket_provider import AWSBraketProvider
-from .braket_job import AWSBraketTask, AWSBraketJob
+from .braket_task import AWSBraketTask, AWSBraketJob
